Create EyeFiDoxieScript.scpt
This AppleScript is used to automatically process images coming from to a Doxie scanner without user intervention.

The Doxie program does not correctly date stamp images, so this does the next best thing. Best used with the Eye-Fi card and a folder action , it will open the Doxie.app and save as a PDF with OCR to a specific folder (e.g. Dropbox, Google Drive).

I am not a professional software developer, so this script could use some refinements.

Enjoy!